third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/preload/document-write/document_write_preload_stubs-expected.txt

This is a testharness.js-based test.
FAIL Accessors like location should be properly stubbed to aid in preloading scripts injected via document.write assert_less_than: expected a number less than 217.24 but got 218.93
Harness: the test ran to completion.
